Creates a whole page storybook
