+++ stories/Page_example.stories.tsx
@@ -1,0 +1,21 @@
+import { Story, Meta } from '@storybook/react'
+
+import { Page, PageProps } from './Page'
+import * as HeaderStories from './Header_example.stories'
+
+export default {
+  title: 'Example/Page',
+  component: Page,
+} as Meta
+
+const Template: Story<PageProps> = (args) => <Page {...args} />
+
+export const LoggedIn = Template.bind({})
+LoggedIn.args = {
+  ...HeaderStories.LoggedIn.args,
+}
+
+export const LoggedOut = Template.bind({})
+LoggedOut.args = {
+  ...HeaderStories.LoggedOut.args,
+}
